TP-223 wip(NotaFiscalTestesUnitarios): adiciona precisão de 100% nos testes unitários do cabeçalho nota fiscal

+++ crud/src/main/java/com/agrotis/trainees/crud/service/ItemNotaFiscalService.java
@@ -27,8 +27,8 @@ public class ItemNotaFiscalService {
         this.produtoService = produtoService;
     }
 
-    private double calcularValorTotal(ItemNotaFiscal item) {
-        double valorTotal = item.getQuantidade() * item.getPreco_unitario();
+    private float calcularValorTotal(ItemNotaFiscal item) {
+        float valorTotal = item.getQuantidade() * item.getPreco_unitario();
         return valorTotal;
     }
